+++ inst/odin/model.R
@@ -27,19 +27,21 @@ eta[2] <- eta_h
 ## Core equations for transitions between compartments:
 
 deriv(U[, ]) <- entrants[i, j] - n_UI[i, j] - exr * U[i, j] +
-  n_AU[i, j] + n_TU[i, j]  + sum(wU[i, j, ]) -
-  sum(n_vbe[i, j, ]) - sum(n_vod[i, j, ]) - sum(n_vos[i, j, ])
+  n_AU[i, j] + n_TU[i, j] - sum(n_diag_rec[i, j, ])  + sum(wU[i, j, ]) -
+  sum(n_vbe[i, j, ]) - sum(n_vod[i, j, ]) - sum(n_vos[i, j, ])  +
+  sum(wdU[i, j, ])
 
-deriv(I[, ]) <- n_UI[i, j] - (sigma + exr) * I[i, j] + sum(wI[i, j, ])
+deriv(I[, ]) <- n_UI[i, j] - (sigma + exr) * I[i, j] + sum(wI[i, j, ]) +
+  sum(wdI[i, j, ])
 
 deriv(A[, ]) <- (1 - (1 - ves[j]) * psi) * sigma * I[i, j] - n_AT[i, j] -
-  n_AU[i, j] - exr * A[i, j] + sum(wA[i, j, ])
+  n_AU[i, j] - exr * A[i, j] + sum(wA[i, j, ]) + sum(wdA[i, j, ])
 
 deriv(S[, ]) <- (1 - ves[j]) * psi * sigma * I[i, j] - n_ST[i, j] -
-  exr * S[i, j] + sum(wS[i, j, ])
+  exr * S[i, j] + sum(wS[i, j, ]) + sum(wdS[i, j, ])
 
 deriv(T[, ]) <- n_ST[i, j] + n_AT[i, j] - exr * T[i, j] - n_TU[i, j] +
-  sum(wT[i, j, ])
+  sum(wT[i, j, ]) +  sum(wdT[i, j, ])
 
 ## Update population size
 N[, ] <- U[i, j] + I[i, j] + A[i, j] + S[i, j] + T[i, j]
@@ -54,12 +56,18 @@ Np[]    <- sum(N[i, ]) * p[i]
 foi_LH[] <- prop_C[i] * Np[i] / sum(Np[])
 lambda[] <- p[i] * beta * (epsilon * prop_C[i] + (1 - epsilon) * sum(foi_LH[]))
 
+
 n_UI[, ]     <- lambda[i] * (1 - vea[j]) * U[i, j]
 n_AT[, ]     <- eta[i] * A[i, j]
 n_AU[, ]     <- nu / (1 - ved[j]) * A[i, j]
 n_ST[, ]     <- mu * S[i, j]
 n_TU[, ]     <- rho * T[i, j]
 screened[, ] <- eta[i] * U[i, j]
+
+# mechanism to record number of times infected by moving diagnosed
+# individuals into stratum with the relevant diagnosis history
+n_diag_rec[, , ] <- diag_rec[i, j, k] * n_TU[i, k]
+
 
 # vaccination
 ## time-varying switch
@@ -68,13 +76,14 @@ vax_switch <- interpolate(vax_t, vax_y, "constant")
 ## Number offered / accepting vaccine
 ## at screening
 n_oos[, , ] <- vos[i, j, k] * screened[i, k] * vax_switch
-n_vos[, , ] <- n_oos[i, j, k] * u[i, j, k]
+n_vos[, , ] <- n_oos[i, j, k] * u_s[i, j, k]
 ## on diagnosis
 n_ood[, , ] <- vod[i, j, k] * n_TU[i, k] * vax_switch
-n_vod[, , ] <- n_ood[i, j, k] * u[i, j, k]
+n_vod[, , ] <- n_ood[i, j, k] * u_d[i, j, k]
 ## on entry - no switch as background rate, adolescent uptake included in vbe
 n_obe[, , ] <- vbe[i, j, k] * entrants[i, k]
 n_vbe[, , ] <- n_obe[i, j, k] * u_vbe
+
 
 # waning
 wU[, , ] <- w[j, k] * U[i, k]
@@ -82,6 +91,17 @@ wI[, , ] <- w[j, k] * I[i, k]
 wA[, , ] <- w[j, k] * A[i, k]
 wS[, , ] <- w[j, k] * S[i, k]
 wT[, , ] <- w[j, k] * T[i, k]
+
+
+# waning diagnosis
+wdU[, , ] <- wd[j, k] * U[i, k]
+wdI[, , ] <- wd[j, k] * I[i, k]
+wdA[, , ] <- wd[j, k] * A[i, k]
+wdS[, , ] <- wd[j, k] * S[i, k]
+wdT[, , ] <- wd[j, k] * T[i, k]
+
+
+
 
 ## outputs
 
@@ -92,8 +112,12 @@ deriv(cum_treated[, ])     <- n_TU[i, j]
 deriv(cum_screened[, ])    <- screened[i, j]
 deriv(cum_offered[, ])     <- n_oos[i, j, j] + n_ood[i, j, j] + n_obe[i, j, j]
 deriv(cum_vaccinated[, ])  <- n_vos[i, j, j] + n_vod[i, j, j] + n_vbe[i, j, j]
+deriv(cum_vaccinated_screen[, ])  <- n_vos[i, j, j]
+
 deriv(cum_vbe[, ])         <- n_vbe[i, j, j]
 deriv(cum_offered_vbe[, ]) <- n_obe[i, j, j]
+deriv(cum_entrants[, ]) <- entrants[i, j]
+
 
 # aggregated time series for fitting mcmc
 output(tot_treated)  <- sum(cum_treated)
@@ -124,8 +148,14 @@ initial(cum_treated[, ])     <- 0
 initial(cum_screened[, ])    <- 0
 initial(cum_offered[, ])     <- 0
 initial(cum_vaccinated[, ])  <- 0
+initial(cum_vaccinated_screen[, ])  <- 0
+
 initial(cum_vbe[, ])         <- 0
 initial(cum_offered_vbe[, ]) <- 0
+
+initial(cum_entrants[, ]) <- 0
+
+
 
 # set up dimensions of compartments
 dim(U) <- c(n_group, n_vax)
@@ -155,6 +185,8 @@ dim(n_ST)     <- c(n_group, n_vax)
 dim(n_TU)     <- c(n_group, n_vax)
 dim(screened) <- c(n_group, n_vax)
 
+
+
 dim(cum_incid)       <- c(n_group, n_vax)
 dim(cum_diag_a)      <- c(n_group, n_vax)
 dim(cum_diag_s)      <- c(n_group, n_vax)
@@ -162,8 +194,18 @@ dim(cum_treated)     <- c(n_group, n_vax)
 dim(cum_screened)    <- c(n_group, n_vax)
 dim(cum_offered)     <- c(n_group, n_vax)
 dim(cum_vaccinated)  <- c(n_group, n_vax)
+dim(cum_vaccinated_screen)  <- c(n_group, n_vax)
+
+
 dim(cum_vbe)         <- c(n_group, n_vax)
 dim(cum_offered_vbe) <- c(n_group, n_vax)
+
+dim(cum_entrants) <- c(n_group, n_vax)
+
+
+dim(n_diag_rec) <- c(n_group, n_vax, n_vax)
+dim(diag_rec)   <- c(n_group, n_vax, n_vax)
+
 
 ## Parameters
 p[]     <- user() # Partner change rate in group L/H
@@ -195,10 +237,19 @@ ves[] <- user() # efficacy against symptoms
 vei[] <- user() # efficacy against infectiousness
 
 u_vbe    <- user() # uptake of VbE
-u[, , ]  <- user() # Uptake matrix for VoD/VoD
+u_d[, , ]  <- user() # Uptake matrix for diagnosis
+u_s[, , ]  <- user() # Uptake matrix for screening
+
 w[, ]    <- user() # Waning map
 vax_t[]  <- user()
 vax_y[]  <- user()
+
+wd[, ]   <- user() # Waning of diagnosis map
+
+diag_rec[, , ] <- user()  ## recording diagnosis history mapping
+
+
+
 
 ## par dimensions
 dim(beta_t)  <- length(tt)
@@ -216,22 +267,42 @@ dim(willing) <- n_vax
 dim(vbe)   <- c(n_group, n_vax, n_vax)
 dim(vod)   <- c(n_group, n_vax, n_vax)
 dim(vos)   <- c(n_group, n_vax, n_vax)
+
 dim(w)     <- c(n_vax, n_vax)
-dim(u)     <- c(n_group, n_vax, n_vax)
+
+
+dim(u_d)     <- c(n_group, n_vax, n_vax)
+dim(u_s)     <- c(n_group, n_vax, n_vax)
+
+
 dim(vax_t) <- user()
 dim(vax_y) <- user()
 
 dim(n_vbe) <- c(n_group, n_vax, n_vax)
 dim(n_vos) <- c(n_group, n_vax, n_vax)
 dim(n_vod) <- c(n_group, n_vax, n_vax)
+
 dim(n_obe) <- c(n_group, n_vax, n_vax)
 dim(n_oos) <- c(n_group, n_vax, n_vax)
 dim(n_ood) <- c(n_group, n_vax, n_vax)
+
+
+
+
 dim(wU)   <- c(n_group, n_vax, n_vax)
 dim(wI)   <- c(n_group, n_vax, n_vax)
 dim(wA)   <- c(n_group, n_vax, n_vax)
 dim(wS)   <- c(n_group, n_vax, n_vax)
 dim(wT)   <- c(n_group, n_vax, n_vax)
+
+
+
+dim(wd) <- c(n_vax, n_vax)
+dim(wdU)   <- c(n_group, n_vax, n_vax)
+dim(wdI)   <- c(n_group, n_vax, n_vax)
+dim(wdA)   <- c(n_group, n_vax, n_vax)
+dim(wdS)   <- c(n_group, n_vax, n_vax)
+dim(wdT)   <- c(n_group, n_vax, n_vax)
 
 output(N)   <- N
 output(lambda) <- lambda