+++ tests/testthat/test-onevax_xvwr.R
@@ -51,11 +51,11 @@ test_that("run_onevax_xvwr works correctly", {
 
   for (i in seq_along(y4e)) {
     # no-one in stratum V is vaccinated again
-  expect_equal(sum(y4e[[i]]$cum_vaccinated[, , 2]), 0)
-  expect_equal(y4e[[i]]$cum_vaccinated[, , c(1, 3)] / primary_uptake[i],
-               y4e[[i]]$cum_screened[, , c(1, 3)] +
-                 y4e[[i]]$cum_treated[, , c(1, 3)])
-  expect_equal(apply(y4e[[i]]$N, 1, sum), rep(6e5, 6), tolerance = 1e-5)
+    expect_equal(sum(y4e[[i]]$cum_vaccinated[, , 2]), 0)
+    expect_equal(y4e[[i]]$cum_vaccinated[, , c(1, 3)] / primary_uptake[i],
+                 y4e[[i]]$cum_screened[, , c(1, 3)] +
+                   y4e[[i]]$cum_treated[, , c(1, 3)])
+    expect_equal(apply(y4e[[i]]$N, 1, sum), rep(6e5, 6), tolerance = 1e-5)
   }
   # check vaccination targeting
   y5e <- run_onevax_xvwr(tt, gp, vea = 1, dur = 1e3, strategy = "VoD(L)+VoA(H)",
@@ -76,7 +76,7 @@ test_that("run_onevax_xvwr works correctly", {
 
   # check length of uptake vector must be 1 or length(gp)
   expect_error(run_onevax_xvwr(tt, gp, vea = 1, dur = 1e3, strategy = "VbE",
-                  primary_uptake = c(0, 0.5, 1)))
+                               primary_uptake = c(0, 0.5, 1)))
   # check length of vea must be 1
   expect_error(run_onevax_xvwr(tt, gp, vea = c(0, 1, 1), dur = 1e3,
                                strategy = "VbE", primary_uptake = 1))
@@ -133,10 +133,11 @@ test_that("run_onevax_xvwr works correctly", {
   # check booster_uptake defaults to primary_uptake
 
   y_prim_only <- run_onevax_xvwr(tt, gp, vea = 0, dur = 1e3,
-         primary_uptake = 0.75, strategy = "VoD(L)+VoA(H)")
+                                 primary_uptake = 0.75,
+                                 strategy = "VoD(L)+VoA(H)")
   y_prim_boost <- run_onevax_xvwr(tt, gp, vea = 0, dur = 1e3,
-        primary_uptake = 0.75, booster_uptake = 0.75,
-                                   strategy = "VoD(L)+VoA(H)")
+                                  primary_uptake = 0.75, booster_uptake = 0.75,
+                                  strategy = "VoD(L)+VoA(H)")
 
   expect_equal(y_prim_only, y_prim_boost)
 })